requirements.txt
PyQt5
moviepy
Pillow
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
watchdog